@@ -1,6 +1,6 @@
 {
   "name": "oidc-client-react",
-  "version": "0.1.0-alpha.5",
+  "version": "0.1.0-alpha.6",
   "description": "A react wrapper for oidc-client",
   "main": "index.js",
   "scripts": {
@@ -36,36 +36,43 @@
   },
   "homepage": "https://github.com/nearmap/oidc-client-react#readme",
   "devDependencies": {
-    "babel-cli": "~6.26.0",
-    "babel-core": "~6.26.0",
-    "babel-eslint": "~8.2.2",
-    "babel-jest": "~22.0.0",
-    "babel-plugin-transform-es2015-modules-umd": "~6.24.1",
-    "babel-preset-env": "~1.6.1",
-    "babel-preset-react": "~6.24.1",
-    "babel-preset-stage-2": "~6.24.1",
-    "enzyme": "~3.3.0",
-    "enzyme-adapter-react-16": "~1.1.1",
-    "eslint": "~4.18.2",
-    "eslint-plugin-babel": "~4.1.2",
-    "eslint-plugin-react": "~7.7.0",
-    "jest": "~22.2.0",
+    "@babel/cli": "^7.0.0-rc.2",
+    "@babel/core": "^7.0.0-rc.2",
+    "@babel/node": "^7.0.0-rc.2",
+    "@babel/plugin-proposal-class-properties": "^7.0.0-rc.2",
+    "@babel/plugin-proposal-decorators": "^7.0.0-rc.2",
+    "@babel/plugin-proposal-numeric-separator": "^7.0.0-rc.2",
+    "@babel/plugin-proposal-optional-catch-binding": "^7.0.0-rc.2",
+    "@babel/plugin-proposal-throw-expressions": "^7.0.0-rc.2",
+    "@babel/plugin-transform-runtime": "^7.0.0-rc.2",
+    "@babel/polyfill": "^7.0.0-rc.2",
+    "@babel/preset-env": "^7.0.0-rc.2",
+    "@babel/preset-react": "^7.0.0-rc.2",
+    "@babel/runtime": "^7.0.0-rc.2",
+    "babel-core": "^7.0.0-bridge.0",
+    "babel-eslint": "^8.0.2",
+    "enzyme": "^3.4.4",
+    "enzyme-adapter-react-16": "^1.2.0",
+    "eslint": "^5.4.0",
+    "eslint-plugin-babel": "^5.1.0",
+    "eslint-plugin-react": "^7.11.1",
+    "jest-cli": "^23.5.0",
     "jwt-decode": "~2.2.0",
     "npm-run-all": "~4.1.2",
     "prop-types": "~15.6.1",
-    "react": "~16.2.0",
-    "react-dom": "~16.2.0",
-    "react-test-renderer": "~16.2.0",
-    "redux": "~3.7.2",
+    "react": "~16.4.0",
+    "react-dom": "~16.4.0",
+    "react-test-renderer": "~16.4.0",
+    "redux": "~4.0.0",
     "remark-cli": "~5.0.0",
     "remark-lint": "~6.0.1"
   },
   "dependencies": {
-    "oidc-client": "~1.4.1"
+    "oidc-client": "~1.5.2"
   },
   "peerDependencies": {
-    "react": "~16.2.0",
-    "react-dom": "~16.2.0",
+    "react": "~16.4.0",
+    "react-dom": "~16.4.0",
     "prop-types": "~15.6.1"
   }
 }
